Move polygon validation call to drawInteractions.vue
