+++ resources/views/videos/show/content.blade.php
@@ -60,6 +60,7 @@
       v-on:attaching-active="handleAttachingLabelActive"
       v-on:swapping-active="handleSwappingLabelActive"
       v-on:seek="seek"
+      v-on:is-invalid-polygon="handleInvalidPolygon"
       ></video-screen>
 <video-timeline
       ref="videoTimeline"
